Create common EtoForms CommandLine project
