+++ sample/THNETII.EtoForms.CmdParserHostedSample/Program.cs
@@ -1,7 +1,6 @@
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Options;
 
 using System;
 using System.CommandLine;
@@ -9,48 +8,16 @@ using System.CommandLine.Builder;
 using System.CommandLine.Hosting;
 using System.CommandLine.Invocation;
 using System.CommandLine.Parsing;
-using System.Threading;
 
 using THNETII.EtoForms.Hosting;
+using THNETII.EtoForms.CommandLine;
 
 namespace THNETII.EtoForms.CmdParserHostedSample
 {
     public static class Program
     {
-        public static ICommandHandler Handler { get; } = CommandHandler.Create(
-        (IHost host, CancellationToken cancelToken) =>
-        {
-            var application = host.Services.GetRequiredService<Eto.Forms.Application>();
-
-            var options = host.Services
-                .GetRequiredService<IOptions<EtoFormsOptions>>().Value;
-            object form = null;
-            if (options.MainForm is Type formType)
-                form = ActivatorUtilities.GetServiceOrCreateInstance(
-                    host.Services, formType);
-
-            using var cancelReg = cancelToken.Register(obj =>
-            {
-                var app = (Eto.Forms.Application)obj;
-                app.Quit();
-            }, application);
-
-            switch (form)
-            {
-                case Eto.Forms.Form mainForm:
-                    application.Run(mainForm);
-                    break;
-                case Eto.Forms.Dialog dialog:
-                    application.Run(dialog);
-                    break;
-                case null:
-                    application.Run();
-                    break;
-            }
-
-            var hostLifetime = host.Services.GetRequiredService<IHostApplicationLifetime>();
-            hostLifetime.StopApplication();
-        });
+        public static ICommandHandler Handler { get; } =
+            EtoFormsCommandHandler.Create();
 
         [STAThread]
         public static int Main(string[] args)
