+++ sample/THNETII.EtoForms.CmdParserHostedSample/THNETII.EtoForms.CmdParserHostedSample.csproj
@@ -6,6 +6,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\THNETII.EtoForms.Hosting\THNETII.EtoForms.Hosting.csproj" />
+    <ProjectReference Include="..\..\src\THNETII.EtoForms.CommandLine\THNETII.EtoForms.CommandLine.csproj" />
   </ItemGroup>
 
   <ItemGroup>
